third_party/WebKit/LayoutTests/http/tests/history/cross-origin-replace-history-object-expected.txt

ALERT: Parent window's history object before attempt to clear: [object History]
ALERT: Attempting to clear parent window's history object:
ALERT: PASS: access to parent.window.history threw an exception.
ALERT: Parent window's history object after attempt to clear: [object History]
